requirements.txt

tkinter
Pillow
socket
qrcode
Flask
pyautogui
numpy
pyaudio
requests
concurrent.futures
